ai-best-practices-example/pyproject.toml
[project]
name = "ai-app"
version = "0.1.0"
description = "AI best practices example: provider-agnostic LLM integration, config-driven prompts, local cluster."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "openai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "tenacity>=8.2.0",
    "chromadb>=0.4.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
